requirements.txt
streamlit==1.37.1
openai==1.42.0
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.3.1
python-docx==1.1.2
markdown==3.6
rank-bm25==0.2.2
requests==2.32.3
PyYAML==6.0.2
tqdm==4.66.4
